requirements.txt
torch==2.2.1
torchvision==0.17.1
Pillow==10.2.0
numpy==1.26.4
chromadb==0.4.22
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
scikit-learn==1.4.1.post1
tqdm==4.66.2
streamlit==1.32.0
requests==2.31.0 